SkillBridgeResume/requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
openpyxl==3.1.2
plotly==5.17.0
scikit-learn==1.3.0
spacy==3.7.2
pytesseract>=0.3.10
Pillow>=10.0.0